requirements.txt
pyyaml
lorem
